ISSUE #478 rename files
