+++ bouncer/src/repo/manipulator/modelconvertor/import/ifcHelper/repo_ifc_utils_geometry.cpp
@@ -19,11 +19,10 @@
 * Allows geometry creation using ifcopenshell
 */
 
-#include <ifcUtils/Ifc2x3.h>
-#include <ifcUtils/Ifc4.h>
+#include <ifcUtils/repo_ifc_utils_ifc2x3.h>
+#include <ifcUtils/repo_ifc_utils_ifc4.h>
 #include "repo_ifc_utils_common.h"
 #include "repo_ifc_utils_geometry.h"
-#include "repo_ifc_utils_constants.h"
 #include "../../../../core/model/bson/repo_bson_factory.h"
 #include <boost/filesystem.hpp>
 #include "../repo_model_import_config_default_values.h"
