+++ ifcUtils/CMakeLists.txt
@@ -6,7 +6,7 @@ endif()
 
 
 foreach(schema ${SCHEMA_VERSIONS})
-	add_library("ifcUtils_${schema}" SHARED ./geometry_handler.cpp ./tree_parser.cpp)
+	add_library("ifcUtils_${schema}" SHARED ./repo_ifc_utils_geometry_handler.cpp ./repo_ifc_utils_tree_parser.cpp)
 	set_target_properties("ifcUtils_${schema}" PROPERTIES COMPILE_FLAGS "-DIFC_UTILS_API_LIBRARY=1 -DIfcSchema=Ifc${schema} -DDEFINED_Ifc${schema}")
 	target_link_libraries("ifcUtils_${schema}" ${Boost_LIBRARIES} ${MONGO_LIBRARIES} ${OCCT_LIBRARIES} ${IFCOPENSHELL_PARSERLIB} ${IFCOPENSHELL_GEOMLIB})
 	install(TARGETS "ifcUtils_${schema}" DESTINATION lib)
